CI: combine into one
only run applist/simple install pages tasks when the update_applist.yml or update-wiki.sh script changes or if the actions is running on a schedule or if a manual repository dispatch is triggered. this skips running that part of the CI on a standard push to other parts of the website.

Combining like this fixes an issue where the website could get built and published before the update applist CI completed if both the .yml/.sh file and another part of the website was modified in one commit/push.